Metrics checking added Yandex and fixed errors
