+++ apps/web/app/layout.tsx
@@ -1,9 +1,7 @@
-'use client';
-
 import { Inter } from 'next/font/google';
 import Link from 'next/link';
-import Script from 'next/script';
 import '@/shared/styles/globals.css';
+import { Analytics } from './Analytics';
 
 const inter = Inter({
   subsets: ['latin', 'cyrillic'],
@@ -22,52 +20,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ru" className={inter.className}>
-      <head>
-        {/* Microsoft Clarity */}
-        <Script id="clarity-script" strategy="afterInteractive">
-          {`
-            (function(c,l,a,r,i,t,y){
-                c[a]=c[a]||function(){(c[a].q=c[a].q||[]).push(arguments)};
-                t=l.createElement(r);t.async=1;t.src="https://www.clarity.ms/tag/"+i;
-                y=l.getElementsByTagName(r)[0];y.parentNode.insertBefore(t,y);
-            })(window, document, "clarity", "script", "sf1va8vj8r");
-        `}
-        </Script>
-        {/* Google Analytics */}
-        <Script
-          strategy="afterInteractive"
-          src="https://www.googletagmanager.com/gtag/js?id=G-3CRXRB8F5K"
-        />
-        <Script id="google-analytics-inline" strategy="afterInteractive">
-          {`
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', 'G-3CRXRB8F5K');
-        `}
-        </Script>
-        {/* Yandex.Metrika counter */}
-        <Script id="yandex-metrika" strategy="afterInteractive">
-          {`
-            (function(m,e,t,r,i,k,a){m[i]=m[i]||function(){(m[i].a=m[i].a||[]).push(arguments)};
-            m[i].l=1*new Date();
-            for (var j = 0; j < document.scripts.length; j++) {if (document.scripts[j].src === r) { return; }}
-            k=e.createElement(t),a=e.getElementsByTagName(t)[0],k.async=1,k.src=r,a.parentNode.insertBefore(k,a)})
-            (window, document, "script", "https://mc.yandex.ru/metrika/tag.js", "ym");
-
-            ym(103343970, "init", {
-                  clickmap:true,
-                  trackLinks:true,
-                  accurateTrackBounce:true,
-                  webvisor:true
-            });
-          `}
-        </Script>
-        <noscript>
-          <div><img src="https://mc.yandex.ru/watch/103343970" style={{position: 'absolute', left: '-9999px'}} alt="" /></div>
-        </noscript>
-      </head>
+      <head></head>
       <body className="min-h-screen flex flex-col bg-gray-50">
+        <Analytics />
         {/* Шапка сайта */}
         <header className="sticky top-0 z-50 w-full bg-white/80 backdrop-blur-md border-b border-gray-200">
           <div className="max-w-6xl mx-auto flex items-center justify-between py-3 px-5">
